Update order-payment.tpl
As you already modified function updateConfirmButton in legal.js this should be the conseqeuence in consideration of the fact that it is NOT necessary to provide a checkbox for confirming terms of service or revocation. 
For legal purposes it's sufficient in my view to display an explanatory text with a message like "Ich erkläre mich mit den AGB einverstanden und bestätige die Widerrufsbelehrung gelesen zu haben." containing links to terms of service and revocation cms pages.